# app/Http/Requests/CarMarkStoreRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

/**
 * @OA\Schema(
 *      required={"name"},
 *      schema="CarMarkStoreRequest",  
 *      @OA\Property(property="name", type="string",  example="kia", description="Модель машины"),
 * )
 */
class CarMarkStoreRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     *
     * @return bool
     */
    public function authorize()
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, mixed>
     */
    public function rules()
    {
        return [
            'name' => 'required|max:20',
        ];
    }
}