# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Carshark",
        "version": "1.0"
    },
    "paths": {
...
        "/api/car_marks": {
            "get": {
                "tags": [
                    "Марки машин"
                ],
                "summary": "Получить список марок машин",
                "description": "Получение списка всех марок машин",
                "operationId": "/api/car_marks(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает список марок машин"
                    }
                }
            },
            "post": {
                "tags": [
                    "Марки машин"
                ],
                "summary": "Добавить марку машины",
                "description": "Создание марки машины",
                "operationId": "/api/car_marks(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarMarkStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Создание новой марки машины"
                    }
                }
            }
        },
        "/api/car_marks/{id}": {
            "get": {
                "tags": [
                    "Марки машин"
                ],
                "summary": "Получить марку машины",
                "description": "Получение марки машины по идентификатору",
                "operationId": "/api/car_marks{car_marks}(GET)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор марки машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает марку машины"
                    }
                }
            },
            "put": {
                "tags": [
                    "Марки машин"
                ],
                "summary": "Редактировать марку машины",
                "description": "Редактирование марки машины по идентификатору",
                "operationId": "/api/car_marks(PUT)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор марки машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarMarkStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Редактирование марки машины"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Марки машин"
                ],
                "summary": "Удалить марку машины",
                "description": "Удаление марки машины по идентификатору",
                "operationId": "/api/car_marks/{car_marks}(DELETE)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор марки машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удаление марки машины"
                    }
                }
            }
        },
        "/api/car_models": {
            "get": {
                "tags": [
                    "Модели машин"
                ],
                "summary": "Получить список моделей машин",
                "description": "Получение списка всех моделей машин",
                "operationId": "/api/car_models(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает список моделей машин"
                    }
                }
            },
            "post": {
                "tags": [
                    "Модели машин"
                ],
                "summary": "Добавить модель машины",
                "description": "Создание модели машины",
                "operationId": "/api/car_models(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarModelStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Создание новой модели машины"
                    }
                }
            }
        },
        "/api/car_models/{id}": {
            "get": {
                "tags": [
                    "Модели машин"
                ],
                "summary": "Получить модель машины",
                "description": "Получение модели машины по идентификатору",
                "operationId": "/api/car_models{car_models}(GET)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор модели машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Возвращает модель машины"
                    }
                }
            },
            "put": {
                "tags": [
                    "Модели машин"
                ],
                "summary": "Редактировать модель машины",
                "description": "Редактирование модели машины по идентификатору",
                "operationId": "/api/car_models(PUT)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор модели машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarModelStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Редактирование модели машины"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Модели машин"
                ],
                "summary": "Удалить модель машины",
                "description": "Удаление модели машины по идентификатору",
                "operationId": "/api/car_models/{car_models}(DELETE)",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор модели машины",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удаление модели машины"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CarMarkStoreRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Модель машины",
                        "type": "string",
                        "example": "kia"
                    }
                },
                "type": "object"
            },
            "CarModelStoreRequest": {
                "required": [
                    "name",
                    "car_mark_id"
                ],
                "properties": {
                    "name": {
                        "description": "Модель машины",
                        "type": "string",
                        "example": "rio"
                    },
                    "car_mark_id": {
                        "description": "Марка машины",
                        "type": "integer",
                        "example": "3"
                    }
                },
                "type": "object"
            },
...
        }
    }
}